refactor: move types to separate files and export type module
- Created `internal_type.ts` to hold internal types.
- Created `type.ts` to hold public types.
- Created `lib.ts` to export all public interfaces.
