+++ internal_type.ts
@@ -1,0 +1,7 @@
+/**
+ * @internal
+ */
+export type UnprintableData = {
+  origWord: string;
+  origNextInput: string;
+};
