+++ lib.ts
@@ -1,0 +1,2 @@
+export * from "./unprintable.ts";
+export type * from "./type.ts";
